Add Playwright additional annotations and Apache License
This commit adds Playwright additional annotations, Apache License, and a Maven `pom.xml` file for the project. It also introduces a new annotation - `LocateBy`, which is employed to specify how to locate an element. Associated code and a `README.md` that guides installation and prerequisites in order to use these annotations are also included.